+++ eth_bridge/ic/src/weth/src/main.rs
@@ -18,44 +18,6 @@ use std::convert::Into;
 use std::iter::FromIterator;
 use std::str::FromStr;
 use std::string::String;
-use types::{Nonce, OutgoingMessage};
-
-mod types;
-
-const TERA_ADDRESS: &str = "timop-6qaaa-aaaab-qaeea-cai";
-const WETH_ADDRESS_ETH: &str = "0x2e130e57021bb4dfb95eb4dd0dd8cfceb936148a";
-
-pub trait ToNat {
-    fn to_nat(&self) -> Nat;
-}
-
-impl ToNat for Principal {
-    fn to_nat(&self) -> Nat {
-        Nat::from(num_bigint::BigUint::from_bytes_be(&self.as_slice()[..]))
-    }
-}
-
-pub trait FromNat {
-    fn from_nat(input: Nat) -> Principal;
-}
-
-impl FromNat for Principal {
-    #[inline(always)]
-    fn from_nat(input: Nat) -> Principal {
-        let be_bytes = input.0.to_bytes_be();
-        let be_bytes_len = be_bytes.len();
-        let padding_bytes = if be_bytes_len > 10 && be_bytes_len < 29 {
-            29 - be_bytes_len
-        } else if be_bytes_len < 10 {
-            10 - be_bytes_len
-        } else {
-            0
-        };
-        let mut p_slice = vec![0u8; padding_bytes];
-        p_slice.extend_from_slice(&be_bytes);
-        Principal::from_slice(&p_slice)
-    }
-}
 
 #[derive(CandidType, Default, Deserialize, Clone)]
 pub struct TxLog {
@@ -357,131 +319,6 @@ async fn approve(spender: Principal, value: Nat) -> TxReceipt {
         TransactionStatus::Succeeded,
     )
     .await
-}
-
-#[update(name = "mint")]
-#[candid_method(update, rename = "mint")]
-async fn mint(nonce: Nonce, payload: Vec<Nat>) -> TxReceipt {
-    let caller = ic::caller();
-    if caller != _get_owner() {
-        return Err(TxError::Unauthorized);
-    }
-    let eth_addr_hex = WETH_ADDRESS_ETH.trim_start_matches("0x");
-    let weth_eth_addr_pid = Principal::from_slice(&hex::decode(eth_addr_hex).unwrap());
-
-    let consume: Result<(bool, String), _> = ic::call(
-        Principal::from_str(TERA_ADDRESS).unwrap(),
-        "consume_message",
-        (&weth_eth_addr_pid, nonce, &payload),
-    )
-    .await;
-
-    if consume.is_ok() {
-        let amount = Nat::from(payload[1].0.clone());
-        let to = Principal::from_nat(payload[0].clone());
-        let to_balance = balance_of(to);
-
-        BALANCES.with(|b| {
-            let mut balances = b.borrow_mut();
-            balances.insert(to, to_balance + amount.clone());
-        });
-        STATS.with(|s| {
-            let mut stats = s.borrow_mut();
-            stats.total_supply += amount.clone();
-        });
-        _history_inc();
-
-        return add_record(
-            caller,
-            Operation::Mint,
-            caller,
-            to,
-            amount,
-            Nat::from(0),
-            ic::time(),
-            TransactionStatus::Succeeded,
-        )
-        .await;
-    }
-
-    Err(TxError::Other(format!(
-        "Consuming message from L1 failed with caller {:?}!",
-        caller
-    )))
-}
-
-#[update(name = "burn")]
-#[candid_method(update, rename = "burn")]
-async fn burn(eth_addr: Principal, amount: Nat) -> TxReceipt {
-    let caller = ic::caller();
-    let caller_balance = balance_of(caller);
-    if caller_balance.clone() < amount.clone() {
-        return Err(TxError::InsufficientBalance);
-    }
-    BALANCES.with(|b| {
-        let mut balances = b.borrow_mut();
-        balances.insert(caller, caller_balance - amount.clone());
-    });
-    STATS.with(|s| {
-        let mut stats = s.borrow_mut();
-        stats.total_supply -= amount.clone();
-    });
-    _history_inc();
-
-    let payload = [eth_addr.clone().to_nat(), amount.clone()];
-    let weth_addr_hex = WETH_ADDRESS_ETH.trim_start_matches("0x");
-    let weth_eth_addr_pid = Principal::from_slice(&hex::decode(weth_addr_hex).unwrap());
-
-    let send_message: Result<(OutgoingMessage, String), _> = ic::call(
-        Principal::from_str(TERA_ADDRESS).unwrap(),
-        "send_message",
-        (&weth_eth_addr_pid, &payload),
-    )
-    .await;
-
-    match send_message {
-        Ok(result) => {
-            let outgoing_message = result.0;
-
-            let msg_hash_as_nat = Nat::from(num_bigint::BigUint::from_bytes_be(
-                &outgoing_message.msg_key,
-            ));
-
-            let add = add_record(
-                caller,
-                Operation::Burn,
-                caller,
-                caller,
-                amount.clone(),
-                Nat::from(0),
-                ic::time(),
-                TransactionStatus::Succeeded,
-            )
-            .await;
-
-            if add.is_ok() {
-                return Ok(msg_hash_as_nat);
-            }
-
-            Err(TxError::LedgerTrap)
-        }
-        Err(_) => {
-            BALANCES.with(|b| {
-                let mut balances = b.borrow_mut();
-                balances.insert(caller, balance_of(caller) + amount.clone());
-            });
-
-            STATS.with(|s| {
-                let mut stats = s.borrow_mut();
-                stats.total_supply += amount.clone();
-            });
-
-            Err(TxError::Other(format!(
-                "Sending message to L1 failed with caller {:?} and {}!",
-                caller, amount
-            )))
-        }
-    }
 }
 
 #[update(name = "setName")]
@@ -810,9 +647,190 @@ async fn insert_into_cap_priv(ie: IndefiniteEvent) -> TxReceipt {
     insert_res
 }
 
+// Proxy
+
+const TERA_ADDRESS: &str = "timop-6qaaa-aaaab-qaeea-cai";
+const WETH_ADDRESS_ETH: &str = "0x2e130e57021bb4dfb95eb4dd0dd8cfceb936148a";
+
+pub type Nonce = Nat;
+
+#[derive(Clone, Debug, CandidType, Deserialize, PartialEq, Eq, Hash)]
+pub struct OutgoingMessage {
+    pub(crate) msg_key: [u8; 32],
+    pub(crate) msg_hash: String,
+}
+
+#[derive(Deserialize, CandidType)]
+pub struct ConsumeMessageParam {
+    pub eth_addr: Principal,
+    pub payload: Vec<Nat>,
+}
+
+#[derive(Deserialize, CandidType)]
+pub struct SendMessageParam {
+    pub eth_addr: Principal,
+    pub payload: Vec<Nat>,
+}
+
+pub trait ToNat {
+    fn to_nat(&self) -> Nat;
+}
+
+impl ToNat for Principal {
+    fn to_nat(&self) -> Nat {
+        Nat::from(num_bigint::BigUint::from_bytes_be(&self.as_slice()[..]))
+    }
+}
+
+pub trait FromNat {
+    fn from_nat(input: Nat) -> Principal;
+}
+
+impl FromNat for Principal {
+    #[inline(always)]
+    fn from_nat(input: Nat) -> Principal {
+        let be_bytes = input.0.to_bytes_be();
+        let be_bytes_len = be_bytes.len();
+        let padding_bytes = if be_bytes_len > 10 && be_bytes_len < 29 {
+            29 - be_bytes_len
+        } else if be_bytes_len < 10 {
+            10 - be_bytes_len
+        } else {
+            0
+        };
+        let mut p_slice = vec![0u8; padding_bytes];
+        p_slice.extend_from_slice(&be_bytes);
+        Principal::from_slice(&p_slice)
+    }
+}
+
+#[update(name = "mint")]
+#[candid_method(update, rename = "mint")]
+async fn mint(nonce: Nonce, payload: Vec<Nat>) -> TxReceipt {
+    let caller = ic::caller();
+    if caller != _get_owner() {
+        return Err(TxError::Unauthorized);
+    }
+    let eth_addr_hex = WETH_ADDRESS_ETH.trim_start_matches("0x");
+    let weth_eth_addr_pid = Principal::from_slice(&hex::decode(eth_addr_hex).unwrap());
+
+    let consume: Result<(bool, String), _> = ic::call(
+        Principal::from_str(TERA_ADDRESS).unwrap(),
+        "consume_message",
+        (&weth_eth_addr_pid, nonce, &payload),
+    )
+    .await;
+
+    if consume.is_ok() {
+        let amount = Nat::from(payload[1].0.clone());
+        let to = Principal::from_nat(payload[0].clone());
+        let to_balance = balance_of(to);
+
+        BALANCES.with(|b| {
+            let mut balances = b.borrow_mut();
+            balances.insert(to, to_balance + amount.clone());
+        });
+        STATS.with(|s| {
+            let mut stats = s.borrow_mut();
+            stats.total_supply += amount.clone();
+        });
+        _history_inc();
+
+        return add_record(
+            caller,
+            Operation::Mint,
+            caller,
+            to,
+            amount,
+            Nat::from(0),
+            ic::time(),
+            TransactionStatus::Succeeded,
+        )
+        .await;
+    }
+
+    Err(TxError::Other(format!(
+        "Consuming message from L1 failed with caller {:?}!",
+        caller
+    )))
+}
+
+#[update(name = "burn")]
+#[candid_method(update, rename = "burn")]
+async fn burn(eth_addr: Principal, amount: Nat) -> TxReceipt {
+    let caller = ic::caller();
+    let caller_balance = balance_of(caller);
+    if caller_balance.clone() < amount.clone() {
+        return Err(TxError::InsufficientBalance);
+    }
+    BALANCES.with(|b| {
+        let mut balances = b.borrow_mut();
+        balances.insert(caller, caller_balance - amount.clone());
+    });
+    STATS.with(|s| {
+        let mut stats = s.borrow_mut();
+        stats.total_supply -= amount.clone();
+    });
+    _history_inc();
+
+    let payload = [eth_addr.clone().to_nat(), amount.clone()];
+    let weth_addr_hex = WETH_ADDRESS_ETH.trim_start_matches("0x");
+    let weth_eth_addr_pid = Principal::from_slice(&hex::decode(weth_addr_hex).unwrap());
+
+    let send_message: Result<(OutgoingMessage, String), _> = ic::call(
+        Principal::from_str(TERA_ADDRESS).unwrap(),
+        "send_message",
+        (&weth_eth_addr_pid, &payload),
+    )
+    .await;
+
+    match send_message {
+        Ok(result) => {
+            let outgoing_message = result.0;
+
+            let msg_hash_as_nat = Nat::from(num_bigint::BigUint::from_bytes_be(
+                &outgoing_message.msg_key,
+            ));
+
+            let add = add_record(
+                caller,
+                Operation::Burn,
+                caller,
+                caller,
+                amount.clone(),
+                Nat::from(0),
+                ic::time(),
+                TransactionStatus::Succeeded,
+            )
+            .await;
+
+            if add.is_ok() {
+                return Ok(msg_hash_as_nat);
+            }
+
+            Err(TxError::LedgerTrap)
+        }
+        Err(_) => {
+            BALANCES.with(|b| {
+                let mut balances = b.borrow_mut();
+                balances.insert(caller, balance_of(caller) + amount.clone());
+            });
+
+            STATS.with(|s| {
+                let mut stats = s.borrow_mut();
+                stats.total_supply += amount.clone();
+            });
+
+            Err(TxError::Other(format!(
+                "Sending message to L1 failed with caller {:?} and {}!",
+                caller, amount
+            )))
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
-    use bigdecimal::BigDecimal;
     use ic_cdk::export::candid::{decode_args, encode_args, Nat};
     use ic_kit::candid::Principal;
     use std::{ops::Mul, str::FromStr};
@@ -865,17 +883,6 @@ mod tests {
         let expected_ether_addr = "f39fd6e51aad88f6f4ce6ab8827279cfffb92266";
         println!("{}", from_principal.to_string());
         assert_eq!(hex::encode(from_principal), expected_ether_addr);
-    }
-
-    #[test]
-    fn test_gwei_to_eth() {
-        let gwei = "0.000000001";
-        let value = BigDecimal::from_str(&"20000000".to_string()).unwrap();
-
-        let result = value.mul(&BigDecimal::from_str(gwei).unwrap());
-        let expected_eth_value = BigDecimal::from_str(&"0.02".to_string()).unwrap();
-
-        assert_eq!(result, expected_eth_value);
     }
 
     #[test]
